requirements/opt-mimetype-magic.txt
# This is an implicit value, here for clarity
--index-url https://pypi.python.org/simple/

file-magic == 0.4.0
